Add nomination page and access rules
