+++ src/pages/AdminPage/MembersTable.tsx
@@ -1,7 +1,6 @@
 import { useState, useEffect, useMemo } from 'react';
 
 import { zodResolver } from '@hookform/resolvers/zod';
-import { InfoCircledIcon } from '@radix-ui/react-icons';
 import { isUserAdmin } from 'lib/users';
 import { SubmitHandler, useController, useForm } from 'react-hook-form';
 import { useQueryClient } from 'react-query';
@@ -17,7 +16,7 @@ import {
 import { zEthAddress } from 'forms/formHelpers';
 import { useApeSnackbar, useApiAdminCircle, useNavigation } from 'hooks';
 import useMobileDetect from 'hooks/useMobileDetect';
-import { Check, X, Slash } from 'icons/__generated';
+import { Check, X, Slash, Info } from 'icons/__generated';
 import { CircleSettingsResult } from 'pages/CircleAdminPage/getCircleSettings';
 import {
   FixedPaymentResult,
@@ -160,7 +159,7 @@ const UserName = ({ user }: { user: IUser }) => {
         {user.name}{' '}
         {user.role === USER_ROLE_COORDINAPE ? (
           <Tooltip content={coordinapeTooltipContent()}>
-            <InfoCircledIcon />
+            <Info size="sm" />
           </Tooltip>
         ) : (
           ''
@@ -437,7 +436,7 @@ const MemberRow = ({
                           </div>
                         }
                       >
-                        <InfoCircledIcon />
+                        <Info size="sm" />
                       </Tooltip>
                     </FormLabel>
                     <CheckBox {...userRole} />
@@ -478,7 +477,7 @@ const MemberRow = ({
                             </div>
                           }
                         >
-                          <InfoCircledIcon />
+                          <Info size="sm" />
                         </Tooltip>
                       </FormLabel>
                       <Flex css={{ gap: '10px', flexWrap: 'wrap' }}>
@@ -536,7 +535,7 @@ const MemberRow = ({
                             </div>
                           }
                         >
-                          <InfoCircledIcon />
+                          <Info size="sm" />
                         </Tooltip>
                       </FormLabel>
                       <Flex css={{ gap: '10px', flexWrap: 'wrap' }}>
