+++ src/routes/paths.ts
@@ -65,6 +65,7 @@ export const paths = {
   history: circlePath('history'),
   members: circlePath('members'),
   membersAdd: circlePath('members/add'),
+  membersNominate: circlePath('members/nominate'),
   team: circlePath('team'),
   vouching: circlePath('vouching'),
   distributions: (circleId: number, epochId: number | string) =>
